+++ src/crypto/hash/cuda/keccak.cu
@@ -19,32 +19,32 @@ extern "C"
 #define KECCAK_STATE_SIZE 25
 #define KECCAK_Q_SIZE 192
 
-__constant__ LONG CUDA_KECCAK_CONSTS[24] = { 0x0000000000000001, 0x0000000000008082,
-                                          0x800000000000808a, 0x8000000080008000, 0x000000000000808b, 0x0000000080000001, 0x8000000080008081,
-                                          0x8000000000008009, 0x000000000000008a, 0x0000000000000088, 0x0000000080008009, 0x000000008000000a,
-                                          0x000000008000808b, 0x800000000000008b, 0x8000000000008089, 0x8000000000008003, 0x8000000000008002,
-                                          0x8000000000000080, 0x000000000000800a, 0x800000008000000a, 0x8000000080008081, 0x8000000000008080,
-                                          0x0000000080000001, 0x8000000080008008 };
+__constant__ uint64_t CUDA_KECCAK_CONSTS[24] = { 0x0000000000000001, 0x0000000000008082,
+0x800000000000808a, 0x8000000080008000, 0x000000000000808b, 0x0000000080000001, 0x8000000080008081,
+0x8000000000008009, 0x000000000000008a, 0x0000000000000088, 0x0000000080008009, 0x000000008000000a,
+0x000000008000808b, 0x800000000000008b, 0x8000000000008089, 0x8000000000008003, 0x8000000000008002,
+0x8000000000000080, 0x000000000000800a, 0x800000008000000a, 0x8000000080008081, 0x8000000000008080,
+0x0000000080000001, 0x8000000080008008 };
 
 typedef struct {
 
-    BYTE sha3_flag;
-    WORD digestbitlen;
-    LONG rate_bits;
-    LONG rate_BYTEs;
-    LONG absorb_round;
+    uint8_t sha3_flag;
+    uint32_t digestbitlen;
+    uint64_t rate_bits;
+    uint64_t rate_BYTEs;
+    uint64_t absorb_round;
 
     int64_t state[KECCAK_STATE_SIZE];
-    BYTE q[KECCAK_Q_SIZE];
+    uint8_t q[KECCAK_Q_SIZE];
 
-    LONG bits_in_queue;
+    uint64_t bits_in_queue;
 
 } cuda_keccak_ctx_t;
 typedef cuda_keccak_ctx_t CUDA_KECCAK_CTX;
 
-__device__ LONG cuda_keccak_leuint64(void *in)
+__device__ uint64_t cuda_keccak_leuint64(void *in)
 {
-    LONG a;
+    uint64_t a;
     memcpy(&a, in, 8);
     return a;
 }
@@ -55,7 +55,7 @@ __device__ int64_t cuda_keccak_MIN(int64_t a, int64_t b)
     return a;
 }
 
-__device__ LONG cuda_keccak_UMIN(LONG a, LONG b)
+__device__ uint64_t cuda_keccak_UMIN(uint64_t a, uint64_t b)
 {
     if (a > b) return b;
     return a;
@@ -63,17 +63,17 @@ __device__ LONG cuda_keccak_UMIN(LONG a, LONG b)
 
 __device__ void cuda_keccak_extract(cuda_keccak_ctx_t *ctx)
 {
-    LONG len = ctx->rate_bits >> 6;
+    uint64_t len = ctx->rate_bits >> 6;
     int64_t a;
-    int s = sizeof(LONG);
+    int32_t s = sizeof(uint64_t);
 
-    for (int i = 0;i < len;i++) {
+    for (int32_t i = 0;i < len;i++) {
         a = cuda_keccak_leuint64((int64_t*)&ctx->state[i]);
         memcpy(ctx->q + (i * s), &a, s);
     }
 }
 
-__device__ __forceinline__ LONG cuda_keccak_ROTL64(LONG a, LONG  b)
+__device__ __forceinline__ uint64_t cuda_keccak_ROTL64(uint64_t a, uint64_t  b)
 {
     return (a << b) | (a >> (64 - b));
 }
@@ -89,7 +89,7 @@ __device__ void cuda_keccak_permutations(cuda_keccak_ctx_t * ctx)
     int64_t *a15 = A + 15, *a16 = A + 16, *a17 = A + 17, *a18 = A + 18, *a19 = A + 19;
     int64_t *a20 = A + 20, *a21 = A + 21, *a22 = A + 22, *a23 = A + 23, *a24 = A + 24;
 
-    for (int i = 0; i < KECCAK_ROUND; i++) {
+    for (int32_t i = 0; i < KECCAK_ROUND; i++) {
 
         /* Theta */
         int64_t c0 = *a00 ^ *a05 ^ *a10 ^ *a15 ^ *a20;
@@ -204,11 +204,11 @@ __device__ void cuda_keccak_permutations(cuda_keccak_ctx_t * ctx)
 }
 
 
-__device__ void cuda_keccak_absorb(cuda_keccak_ctx_t *ctx, BYTE* in)
+__device__ void cuda_keccak_absorb(cuda_keccak_ctx_t *ctx, uint8_t* in)
 {
 
-    LONG offset = 0;
-    for (LONG i = 0; i < ctx->absorb_round; ++i) {
+    uint64_t offset = 0;
+    for (uint64_t i = 0; i < ctx->absorb_round; ++i) {
         ctx->state[i] ^= cuda_keccak_leuint64(in + offset);
         offset += 8;
     }
@@ -218,24 +218,24 @@ __device__ void cuda_keccak_absorb(cuda_keccak_ctx_t *ctx, BYTE* in)
 
 __device__ void cuda_keccak_pad(cuda_keccak_ctx_t *ctx)
 {
-    ctx->q[ctx->bits_in_queue >> 3] |= (1L << (ctx->bits_in_queue & 7));
+    ctx->q[ctx->bits_in_queue >> 3] |= (1LL << (ctx->bits_in_queue & 7));
 
     if (++(ctx->bits_in_queue) == ctx->rate_bits) {
         cuda_keccak_absorb(ctx, ctx->q);
         ctx->bits_in_queue = 0;
     }
 
-    LONG full = ctx->bits_in_queue >> 6;
-    LONG partial = ctx->bits_in_queue & 63;
+    uint64_t full = ctx->bits_in_queue >> 6;
+    uint64_t partial = ctx->bits_in_queue & 63;
 
-    LONG offset = 0;
-    for (int i = 0; i < full; ++i) {
+    uint64_t offset = 0;
+    for (int32_t i = 0; i < full; ++i) {
         ctx->state[i] ^= cuda_keccak_leuint64(ctx->q + offset);
         offset += 8;
     }
 
     if (partial > 0) {
-        LONG mask = (1L << partial) - 1;
+        uint64_t mask = (1LL << partial) - 1;
         ctx->state[full] ^= cuda_keccak_leuint64(ctx->q + offset) & mask;
     }
 
@@ -250,7 +250,7 @@ __device__ void cuda_keccak_pad(cuda_keccak_ctx_t *ctx)
 /*
  * Digestbitlen must be 128 224 256 288 384 512
  */
-__device__ void cuda_keccak_init(cuda_keccak_ctx_t *ctx, WORD digestbitlen)
+__device__ void cuda_keccak_init(cuda_keccak_ctx_t *ctx, uint32_t digestbitlen)
 {
     memset(ctx, 0, sizeof(cuda_keccak_ctx_t));
     ctx->sha3_flag = 0;
@@ -264,13 +264,13 @@ __device__ void cuda_keccak_init(cuda_keccak_ctx_t *ctx, WORD digestbitlen)
 /*
  * Digestbitlen must be 224 256 384 512
  */
-__device__ void cuda_keccak_sha3_init(cuda_keccak_ctx_t *ctx, WORD digestbitlen)
+__device__ void cuda_keccak_sha3_init(cuda_keccak_ctx_t *ctx, uint32_t digestbitlen)
 {
     cuda_keccak_init(ctx, digestbitlen);
     ctx->sha3_flag = 1;
 }
 
-__device__ void cuda_keccak_update(cuda_keccak_ctx_t *ctx, BYTE *in, LONG inlen)
+__device__ void cuda_keccak_update(cuda_keccak_ctx_t *ctx, uint8_t *in, uint64_t inlen)
 {
     int64_t BYTEs = ctx->bits_in_queue >> 3;
     int64_t count = 0;
@@ -296,16 +296,16 @@ __device__ void cuda_keccak_update(cuda_keccak_ctx_t *ctx, BYTE *in, LONG inlen)
     ctx->bits_in_queue = BYTEs << 3;
 }
 
-__device__ void cuda_keccak_final(cuda_keccak_ctx_t *ctx, BYTE *out)
+__device__ void cuda_keccak_final(cuda_keccak_ctx_t *ctx, uint8_t *out)
 {
     if (ctx->sha3_flag) {
-        int mask = (1 << 2) - 1;
-        ctx->q[ctx->bits_in_queue >> 3] = (BYTE)(0x02 & mask);
+        int32_t mask = (1 << 2) - 1;
+        ctx->q[ctx->bits_in_queue >> 3] = (uint8_t)(0x02 & mask);
         ctx->bits_in_queue += 2;
     }
 
     cuda_keccak_pad(ctx);
-    LONG i = 0;
+    uint64_t i = 0;
 
     while (i < ctx->digestbitlen) {
         if (ctx->bits_in_queue == 0) {
@@ -314,22 +314,22 @@ __device__ void cuda_keccak_final(cuda_keccak_ctx_t *ctx, BYTE *out)
             ctx->bits_in_queue = ctx->rate_bits;
         }
 
-        LONG partial_block = cuda_keccak_UMIN(ctx->bits_in_queue, ctx->digestbitlen - i);
+        uint64_t partial_block = cuda_keccak_UMIN(ctx->bits_in_queue, ctx->digestbitlen - i);
         memcpy(out + (i >> 3), ctx->q + (ctx->rate_BYTEs - (ctx->bits_in_queue >> 3)), partial_block >> 3);
         ctx->bits_in_queue -= partial_block;
         i += partial_block;
     }
 }
 
-__global__ void kernel_keccak_hash(BYTE* indata, WORD inlen, BYTE* outdata, WORD n_batch, WORD KECCAK_BLOCK_SIZE)
+__global__ void kernel_keccak_hash(uint8_t* indata, uint32_t inlen, uint8_t* outdata, uint32_t n_batch, uint32_t KECCAK_BLOCK_SIZE)
 {
-    WORD thread = blockIdx.x * blockDim.x + threadIdx.x;
+    uint32_t thread = blockIdx.x * blockDim.x + threadIdx.x;
     if (thread >= n_batch)
     {
         return;
     }
-    BYTE* in = indata  + thread * inlen;
-    BYTE* out = outdata  + thread * KECCAK_BLOCK_SIZE;
+    uint8_t* in = indata  + thread * inlen;
+    uint8_t* out = outdata  + thread * KECCAK_BLOCK_SIZE;
     CUDA_KECCAK_CTX ctx;
     cuda_keccak_init(&ctx, KECCAK_BLOCK_SIZE << 3);
     cuda_keccak_update(&ctx, in, inlen);
@@ -337,17 +337,17 @@ __global__ void kernel_keccak_hash(BYTE* indata, WORD inlen, BYTE* outdata, WORD
 }
 extern "C"
 {
-void mcm_cuda_keccak_hash_batch(BYTE * in, WORD inlen, BYTE * out, WORD n_outbit, WORD n_batch)
+void mcm_cuda_keccak_hash_batch(uint8_t * in, uint32_t inlen, uint8_t * out, uint32_t n_outbit, uint32_t n_batch)
 {
-    BYTE * cuda_indata;
-    BYTE * cuda_outdata;
-    const WORD KECCAK_BLOCK_SIZE = (n_outbit >> 3);
+    uint8_t * cuda_indata;
+    uint8_t * cuda_outdata;
+    const uint32_t KECCAK_BLOCK_SIZE = (n_outbit >> 3);
     cudaMalloc(&cuda_indata, inlen * n_batch);
     cudaMalloc(&cuda_outdata, KECCAK_BLOCK_SIZE * n_batch);
     cudaMemcpy(cuda_indata, in, inlen * n_batch, cudaMemcpyHostToDevice);
 
-    WORD thread = 256;
-    WORD block = (n_batch + thread - 1) / thread;
+    uint32_t thread = 256;
+    uint32_t block = (n_batch + thread - 1) / thread;
 
     kernel_keccak_hash << < block, thread >> > (cuda_indata, inlen, cuda_outdata, n_batch, KECCAK_BLOCK_SIZE);
     cudaMemcpy(out, cuda_outdata, KECCAK_BLOCK_SIZE * n_batch, cudaMemcpyDeviceToHost);
